backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
timescaledb-psycopg2==0.1.3

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
pyjwt==2.8.0
passlib[argon2]==1.7.4
authlib==1.2.1
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# Task Queue
celery==5.3.4
flower==2.0.1

# Machine Learning & Data Science
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
prophet==1.1.4
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# NLP & AI
spacy==3.7.2
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
openai==1.3.7
whisper-openai==1.1.10

# API & HTTP
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# Testing (Optional - use requirements-dev.txt for dev)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# httpx==0.25.2

# Email
fastapi-mail==1.4.1

# File Processing
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Rate Limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0

# JSON
orjson==3.9.10

# Time Series
statsmodels==0.14.0

# Visualization (for ML training)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Excel Processing
xlrd==2.0.1
XlsxWriter==3.1.9